+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWGDN7FQ7NBKZVVRBG0EM1VY"
+        "01JWQSTP5ABT64EWVNNSV5KQZT"
     )
     r"""The id of the resource"""