+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-30T11:01:23.565Z")
+    created: Optional[datetime] = parse_datetime("2025-06-02T07:48:46.107Z")
     r"""The date and time the resource was created"""