+++ packages/orq-rc/RELEASES.md
@@ -1599,3 +1599,13 @@ Based on:
 - [python v3.7.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v3.7.0-rc.2] https://pypi.org/project/orq-ai-sdk/3.7.0-rc.2 - packages/orq-rc
+
+## 2025-06-02 07:50:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.554.1 (2.616.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.7.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v3.7.0-rc.3] https://pypi.org/project/orq-ai-sdk/3.7.0-rc.3 - packages/orq-rc